+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 * v5.0.2 in progress
+  * Address [#48](https://github.com/clj-holmes/clj-watson/issues/48) by updating all of the project dependencies, including DependencyCheck to 9.0.8.
   * Address [#47](https://github.com/clj-holmes/clj-watson/issues/47) by printing out the optional properties read from the `clj-watson.properties` file.
   * Documentation improvements.